+++ src/main/java/com/tecpro/guiacultural/crud/CRUDVisualArts.java
@@ -16,23 +16,29 @@ import org.javalite.activejdbc.Model;
  * @author joako
  */
 public class CRUDVisualArts extends CRUDEvent {
+
     public VisualArts create(String name, String location, float price, String date_from, String date_until, String hour_from, String hour_until,
             int organizer_id, String description, String artist) {
         openBase();
         Base.openTransaction();
-        VisualArts visualArts = VisualArts.createIt("name", name,
-                "location",location,
-                "price",price,
-                "date_from",date_from,
-                "date_until",date_until,
-                "hour_from",hour_from,
-                "hour_until",hour_until,
-                "description",description,
-                "artist",artist);
-        Organizer organizer = Organizer.findById(organizer_id);
-        organizer.add(visualArts);
-        Base.commitTransaction();
-        return visualArts;
+        VisualArts visualArts = VisualArts.create("name", name,
+                "location", location,
+                "price", price,
+                "date_from", date_from,
+                "date_until", date_until,
+                "hour_from", hour_from,
+                "hour_until", hour_until,
+                "description", description,
+                "artist", artist);
+        if (visualArts.save()) {
+            Organizer organizer = Organizer.findById(organizer_id);
+            if (organizer != null) {
+                organizer.add(visualArts);
+                Base.commitTransaction();
+                return visualArts;
+            }
+        }
+        return null;
     }
 
     public VisualArts get(int id) {
@@ -50,8 +56,11 @@ public class CRUDVisualArts extends CRUDEvent {
     public boolean delete(int id) {
         openBase();
         Base.openTransaction();
+        boolean result = false;
         VisualArts visualArts = VisualArts.findById(id);
-        boolean result = visualArts.delete();
+        if (visualArts != null) {
+            result = visualArts.delete();
+        }
         Base.commitTransaction();
         return result;
     }
@@ -61,12 +70,13 @@ public class CRUDVisualArts extends CRUDEvent {
         openBase();
         Base.openTransaction();
         VisualArts visualArts = VisualArts.findById(id);
-        visualArts.set("name",name,"location",location,"price",price,"date_from",date_from,"date_until",date_until,"hour_from",
-                hour_from,"hour_until",hour_until,"description",description,"artist",artist);
-        visualArts.saveIt();
+        if (visualArts != null) {
+            visualArts.set("name", name, "location", location, "price", price, "date_from", date_from, "date_until", date_until, "hour_from",
+                    hour_from, "hour_until", hour_until, "description", description, "artist", artist);
+            visualArts.save();
+        }
         Base.commitTransaction();
         return visualArts;
     }
 
-    
 }
